+++ build-logic/convention/build.gradle.kts
@@ -11,12 +11,17 @@ gradlePlugin {
     plugins {
         register("app-base-setting") {
             id = "app-base-setting"
-            implementationClass = "TrainAppSetting"
+            implementationClass = "TrainAppPlugin"
         }
 
         register("library-common-setting") {
             id = "library-common-setting"
-            implementationClass = "TrainAppLibraryCommonSetting"
+            implementationClass = "TrainAppLibraryCommonPlugin"
+        }
+
+        register("ktlint-setting") {
+            id = "ktlint-setting"
+            implementationClass= "KtlintPlugin"
         }
     }
 }
