+++ build-logic/convention/src/main/kotlin/TrainAppLibraryCommonPlugin.kt
@@ -3,7 +3,7 @@ import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
 
-class TrainAppLibraryCommonSetting : Plugin<Project> {
+class TrainAppLibraryCommonPlugin : Plugin<Project> {
 
     override fun apply(target: Project) {
         with(target) {
