+++ build-logic/convention/src/main/kotlin/TrainAppPlugin.kt
@@ -1,12 +1,9 @@
 import com.android.build.api.dsl.ApplicationExtension
-import com.android.build.gradle.AppExtension
-import com.android.build.gradle.LibraryExtension
-import org.gradle.api.JavaVersion
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
 
-class TrainAppSetting : Plugin<Project> {
+class TrainAppPlugin : Plugin<Project> {
 
     override fun apply(target: Project) {
         with(target) {
